chore(ChangeLogSettings): Use full-length SHA1s by default
Sometimes GitHub is not able to automatically resolve short SHA1s to
commits as part of its release notes, so use full SHA1s instead; they
will only be rendered as short SHA1s as part of notes attched to GitHub
releases anyway.

This also turns `ChangeLogSettings` into a data class in order to
conveniently be able to `copy()` it to alter only selected properties.